backend/requirements.txt
django-ninja==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1